adding backend interface and title component
