+++ src/UI/backend/backend_interface.py
@@ -10,3 +10,19 @@ class LanguageAgent:
             response_parts.append(f"📁 File received: {file_name}") 
         response_parts.append(f"{reversed_words}") 
         return {"response": "\n".join(response_parts)}
+
+#This file provides an abstraction from the backend logic
+class BackendInterface:
+    def __init__(self):
+        self.agent = LanguageAgent()
+
+    def get_agent_response(self, prompt, file_text="", file_name=""):
+        return self.agent.process(prompt, file_text, file_name)
+
+
+#echoing input
+
+
+def get_simulated_response(user_input: str) -> dict:
+    agent = LanguageAgent()
+    return agent.process(user_input)
