+++ src/UI/components/title.py
@@ -1,0 +1,4 @@
+import streamlit as st
+
+def render_title():
+    st.title("🤖 SLAM")
